requirements.txt
torch
torchvision
scikit-learn
numpy
Pillow
tqdm
